Phase 6 PR5 (Phase B): PalaceCatalog — CatalogDomain + OPDS2 data + DI seams
Continues PalaceCatalog SPM extraction. Adds CatalogDomain (CatalogAPI,
CatalogModels, CatalogRepository, OPDSParser), TPPProblemDocument,
TPPOpenSearchDescription, String+TPPStringAdditions, SearchFormatEntry,
7 OPDS2 data files, OPDSFormat enum, and pure-data OPDS2 publication
types into the package. Adds PalaceNetwork as a package dependency.

**Structural design — DI seams replace singleton reaches:**

- New `FeatureFlagProvider` protocol in package. `DefaultCatalogAPI` now
  takes it via init instead of reading `RemoteFeatureFlags.shared`. App
  target wires up `RemoteFeatureFlags` via empty conformance extension.
- `TPPOpenSearchDescription.withURL` now takes a `NetworkClient`
  parameter (from PalaceNetwork) instead of reaching for
  `AppContainer.production().networkExecutor`. The static method is
  back inside the type — no orphan +Networking extension.
- `TPPProblemDocument.forExpiredOrMissingCredentials` (uses Strings) and
  `TPPProblemDocument.fromResponseError` (uses NSError extension) split
  to `Palace/ErrorHandling/TPPProblemDocument+Localized.swift` in app
  target.

**Files moved into package (14 new this phase):**

- CatalogDomain/{CatalogAPI, CatalogModels, CatalogRepository, OPDSParser}
- TPPProblemDocument, TPPOpenSearchDescription, String+TPPStringAdditions
- 7 pure OPDS2 data types: AuthenticationDocument, CatalogsFeed, Feed,
  Link, LinkArray, Publication
- New: FeatureFlagProvider.swift (DI seam)
- New: OPDS2PublicationData.swift (extracted pure-data structs from
  OPDS2PublicationExtended.swift; the TPPBook bridge stays in app)
- New: OPDSFormat.swift (extracted from UnifiedOPDSService.swift; the
  AppContainer-coupled UnifiedOPDSService stays in app)

**Files staying in app target (TPPBook bridge / AppContainer reads):**

- OPDS2/Models/OPDS2PublicationExtended.swift (toBook conversions)
- OPDS2/OPDSFeedService.swift (TPPBook construction)
- OPDS2/Service/UnifiedOPDSService.swift (AccountsManager reads)
- OPDS2/Cache/OPDSFeedCache.swift (uses GeneralCache from ImageCache)
- New: TPPProblemDocument+Localized.swift (Strings + NSError ext)

**Scope:** Phase B only — green standalone package build. App-target
build will fail at this commit until phase E (Xcode wiring +
DefaultCatalogAPI consumer updates + RemoteFeatureFlags conformance).

**Not done:**
- Tests not yet updated for new DefaultCatalogAPI init signature (4 sites
  in PalaceTests/) — phase E's TDD work
- 4 app-target DefaultCatalogAPI() construction sites need updating
- RemoteFeatureFlags: FeatureFlagProvider conformance not yet declared
- Xcode project not yet wired to the package
- Phase E (consumer updates + Xcode wiring) and F (tests + ForgeOS gates
  + PR) pending

**Deferred:**
- Phase 7 god-class decomp (MyBooksDownloadCenter)
- Other SPM extractions (Auth, Reader, Audiobooks)
- TPPBook protocol abstraction (deferred — bridges stay in app target)

ForgeOS changeset: cs_bd911419
