+++ Palace/Packages/PalaceCatalog/Sources/PalaceCatalog/CatalogModels.swift
@@ -78,6 +78,27 @@ public struct CatalogEntry: Identifiable {
     }
 }
 
-// SearchFormatEntry is defined in CatalogUI/ViewModels/CatalogFilterModels.swift
-// (kept there because the ViewModel layer was the original consumer; the
-// CatalogDomain layer references it via this comment for clarity).
+/// A format entry point shown in the search screen filter row.
+/// Extracted from the groups feed's entry-point facets (e.g. All, eBooks, Audiobooks).
+public struct SearchFormatEntry: Identifiable, Hashable {
+    public let id: String
+    public let title: String
+
+    /// Groups feed URL for this format (e.g. /groups/?entrypoint=Book).
+    /// Used to lazily fetch the format-specific search descriptor URL.
+    public let groupsFeedURL: URL
+
+    /// OpenSearch descriptor URL for this format.
+    /// Populated immediately for the active format; nil for others until first use.
+    public let searchDescriptorURL: URL?
+
+    public let isActive: Bool
+
+    public init(id: String, title: String, groupsFeedURL: URL, searchDescriptorURL: URL?, isActive: Bool) {
+        self.id = id
+        self.title = title
+        self.groupsFeedURL = groupsFeedURL
+        self.searchDescriptorURL = searchDescriptorURL
+        self.isActive = isActive
+    }
+}
