+++ Palace/Packages/PalaceCatalog/Sources/PalaceCatalog/OPDS2AuthenticationDocument.swift
@@ -8,16 +8,16 @@
 
 import Foundation
 
-enum OPDS2LinkRel: String {
+public enum OPDS2LinkRel: String {
     case passwordReset = "http://librarysimplified.org/terms/rel/patron-password-reset"
 }
 
-struct Announcement: Codable {
+public struct Announcement: Codable {
     let id: String
     let content: String
 }
 
-struct OPDS2AuthenticationDocument: Codable {
+public struct OPDS2AuthenticationDocument: Codable {
     struct Features: Codable {
         let disabled: [String]?
         let enabled: [String]?
