+++ Palace/Packages/PalaceCatalog/Sources/PalaceCatalog/OPDS2LinkArray.swift
